:truck: Replaced metrics.py <- metrics_abdiel.py
The files were almost identical, this is my branch and the purpose of that file was to include the new metrics defined by Abdiel on the experiments done on my branch.
